+++ corsafe-poc-application/src/store.js
@@ -7,7 +7,7 @@ import registerReducer from './reducers/registerReducer';
 import loginReducer from './reducers/loginReducer';
 const rootReducer = combineReducers({
   register: registerReducer,
-  login: loginReducer,
+  auth: loginReducer,
 });
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
